Add DataEncoding and KeyEncoding.
